Move extra-optimizers to separate module
