+++ src/torch_constrained/__init__.py
@@ -1,1 +1,2 @@
 from .constrained_optimizer import *
+from .extra_optimizers import *
